Adds a --sources-url CLI option to fetch sources from additional URLs.
This PR adds a --sources-url CLI option. It is a comma separated list of
URLs that EasyBuild will fetch sources from. It replaces the hard-coded
EASYBUILD_SOURCES_URL, but keeps it as a default value. The --sources-url-priority
option is used to determine if these URLs are to be queried first, or last, i.e.
as a fallback.